Flytt fødselsdato-kode inn til andre dato-utils
